Create Detect cycle in a directed graph using BFS
Given a graph... detect if it has cycle or not using BFS. 

To solve this problem, we used solution of Returning topological sort of a graph using BFS . Here we used a count which increases after Dequeuing each vertex . Then we compare the count with total vertex number. If (count != total_vertex) then it means the graph has a cycle and if (count == total_vertex) then it means the graph has no cycle and each vertex was visited.