Create Bipartite Graph
Check if a graph is bipartite or not. If a graph can be divided into 2 groups such that none of its' node has the same color as its neighbour then it's a Bipartite graph. If a node has no option but to have the same color as its' neighbour then it's not a Bipartite graph.

Only graphs with ODD node cycle can not be a Bipartite Graph.